Update 2020-09-02-KJEZ35-custom-turtle
So to begin I started by following the example that was given on the materials page for the custom turtle. I actually tried to follow the instructions but more or less build it myself. However, one issue that continue to arise was that I could not get the background to change colors. I tried different commands along with adding words and colors but nothing seemed to work. I was able to change the color of the turtle. While I enjoyed the process of playing with the code to make the colors appear, it was very frustrating dealing with the background. For the next assignment I simply went to the tabs on trinket that allowed for me to copy the link and added as instructed. One thing I will have refer back to is how to know if it the copy was successful. I also added a link to show an image because I was not sure if that would work as oppose to the entire trinket platform.